requirements.txt
Flask
gunicorn
tensorflow
opencv-python
numpy
pillow
sqlite3-binary
